% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{allele_stats}
\alias{allele_stats}
\title{Gets allele stats}
\usage{
allele_stats(data, converted)
}
\arguments{
\item{data}{A data.frame.}

\item{converted}{Specify the list column containing the converted allele calls (output from convert_rel)}
}
\description{
Gets allele stats
}
