% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{allele_stats_}
\alias{allele_stats_}
\title{Gets allele stats}
\usage{
allele_stats_(data, converted_)
}
\arguments{
\item{data}{A data.frame.}

\item{converted_}{String defining the list column containing the converted allele calls (output from convert_rel)}
}
\description{
Gets allele stats
}

