% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constructors.R
\name{asmap_prog}
\alias{asmap_prog}
\title{Runs mstmap for a linkage groups of map objects that
satisfy threshold criteria along sequence of p.values.
Intermediate results are kept to trace the point at which
linkage groups break.}
\usage{
asmap_prog(obj, comp, p.values = 10^-(c(5:10)), redo_threshold = 20)
}
\arguments{
\item{obj}{A cross objects suitable for mstmap.}

\item{comp}{A comparison object - currently unused}

\item{p.values}{A decreasing vector of p.values between 0 and 1.}

\item{redo_threshold}{Distance in cM. If a linkage group has a
gap large than this threshold, it will be recalculated in the next cycle.}
}
\description{
Runs mstmap for a linkage groups of map objects that
satisfy threshold criteria along sequence of p.values.
Intermediate results are kept to trace the point at which
linkage groups break.
}

