% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constructors.R
\name{asmap_step}
\alias{asmap_step}
\title{A wrapper around ASMap::mstmap with some error handling.
Associates success or error and p.value with the output.}
\usage{
asmap_step(obj, p.value = 1e-06)
}
\arguments{
\item{obj}{A cross object suitable for mstmap().}

\item{p.value}{A value between 0 and 1.}
}
\description{
A wrapper around ASMap::mstmap with some error handling.
Associates success or error and p.value with the output.
}
