% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{bind_map_list}
\alias{bind_map_list}
\title{Bind a list of mapdfs together}
\usage{
bind_map_list(mapdf_list)
}
\arguments{
\item{mapdf_list}{A list of map dfs}
}
\value{
A tidy_gen_map
}
\description{
Bind a list of mapdfs together
}
