% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bp_cross}
\alias{bp_cross}
\title{A biparental cross object}
\format{A list of class 'cross' produced by the qtl package
\describe{
\item{geno}{A list of the genotyping data}
\item{map}{A list describing the genetic map}
}}
\usage{
bp_cross
}
\description{
A cross produced by qtl
}
\keyword{datasets}
