% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bp_map}
\alias{bp_map}
\title{A biparental map object}
\format{A list of class 'map'
\describe{
\item{1}{A vector of class 'A' describing a linkage group}
...
}}
\usage{
bp_map
}
\description{
A map object produced by qtl. This is the same object as
genomap::bp_cross$map
}
\keyword{datasets}
