% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{check_parents_}
\alias{check_parents_}
\title{Checks informativeness of parents}
\usage{
check_parents_(data, maternal, paternal, missingString = "--")
}
\arguments{
\item{data}{A data.frame where each row is a marker score}

\item{maternal, paternal}{Strings giving names of maternal, paternal columns}

\item{missingString}{A string defining missing scores}
}
\description{
Checks informativeness of parents
}
