% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{convert_rel}
\alias{convert_rel}
\title{Convert absolute scores to relative}
\usage{
convert_rel(data, markerName, maternal, paternal, ..., missingString = "--")
}
\arguments{
\item{data}{A data frame}

\item{markerName}{Specification of the column containing the marker name}

\item{maternal}{Specification of the column containing the maternal score}

\item{paternal}{Specification of the column containing the paternal score}

\item{...}{Specification of the columns containing progeny scores}

\item{missingString}{A string representing a missing score}
}
\description{
Works one marker at a time. Takes absolute genotype scores and converts to relative scores (AA/AB/BB to
match maternal/het/paternal).
}
