% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{convert_rel_}
\alias{convert_rel_}
\title{Convert absolute scores to relative (standard-evaluation)}
\usage{
convert_rel_(data, markerName_, maternal_, paternal_, progeny_cols_,
  missingString = "--")
}
\arguments{
\item{data}{A data frame}

\item{markerName_, maternal_, paternal_}{Strings giving names of marker
maternal and paternal columns}

\item{progeny_cols_}{Character vector giving column names of progeny scores}

\item{missingString}{A string representing a missing score}
}
\description{
This is a S3 generic.
}
\keyword{internal}

