% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{convertScore}
\alias{convertScore}
\title{Convert absolute scores to relative}
\usage{
convertScore(maternal, paternal, progeny, markerName = "unknown",
  missingString = "--")
}
\arguments{
\item{maternal}{The maternal score}

\item{paternal}{The paternal score}

\item{progeny}{A vector of progeny scores}

\item{markerName}{The name of the marker that the scores are for}

\item{missingString}{A string representing a missing score}
}
\description{
Takes absolute genotype scores and converts to relative scores (AA/AB/BB to
match maternal/het/paternal).
}

