% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{geno2df}
\alias{geno2df}
\title{Convert R/qtl genotyping data into data.frame}
\usage{
geno2df(map)
}
\arguments{
\item{map}{An R/qtl cross object}
}
\description{
Similar to map2df, but for the genotying values.
}

