% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotters.R
\name{genoSeg}
\alias{genoSeg}
\title{Compares segregation of markers based on agreement of two markers (population parents)}
\usage{
genoSeg(df, genos, parent1, parent2, missingString = "--")
}
\arguments{
\item{df}{A data.frame}

\item{genos}{Not currently used}

\item{parent1}{The first parent (usually female) - an unquoted string}

\item{parent2}{The second parent (usually male) - an unquoted string}

\item{missingString}{A string value defining missing genotype scores}
}
\description{
Compares segregation of markers based on agreement of two markers (population parents)
}

