% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genosegDH.R
\name{genosegDH}
\alias{genosegDH}
\title{Function for segmentation}
\usage{
genosegDH(geno, genocols, chromosomes, chrname = "chr", posname = "pos", 
    output.common = "seg", data.type = "logratio", allele1.name = "A", 
    allele2.name = "B", missing.name = c("0"), allele1.code = -1, allele2.code = 1, 
    min.seg.size = 1e+05, cna.alpha = 0.01, cna.nperm = 10000, cna.p.method = "perm", 
    cna.eta = 0.01, cna.min.width = 5, seg.mean.cutoffs = c(-0.8, 0.8))
}
\arguments{
\item{geno}{genotyping data, containing at least chr; pos; genotyping scores of lines}

\item{genocols}{column numbers of individuals}

\item{chromosomes}{chromosomes or contigs of interest for segmentation}

\item{chrname}{column name/num of chromosomes or contigs}

\item{posname}{column name/num of marker positions}

\item{output.common}{the prefix name of output file}

\item{data.type}{binary or logratio, default logratio}

\item{missing.name}{code for missing data, could be a vector}

\item{min.seg.size}{minimal size of segments, segments with smaller length will be ignored}

\item{seg.mean.cutoffs}{a numeric vector with two values, c(smaller, larger); default=c(-0.8, 0.8);
default values correspond -1 and 1 as input values of allele1.code and allele2.code,respectively.
          segment mean values <= smaller: allele 1
          segment mean values >= larger: allele 2}

\item{alleleX.name}{allele labels in the output}

\item{alleleX.code}{allele codes as values for segmentation}

\item{cna.xxx}{are DNAcopy parameters, please refer DNAcopy mannual}
}
\value{
output file and variable of the genotyping segmentation result
}
\description{
DH segmentation using genotyping data
Input data requirment: genotyping data with columns of chr; pos; >=1 lines
The package DNAcopy is required.
}
\examples{
genosegDH(geno = "")

}
\seealso{
\url{https://bioconductor.org/packages/release/bioc/html/DNAcopy.html}
}
\author{
Sanzhen Liu
}
