% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genosegF2.R
\name{genosegF2}
\alias{genosegF2}
\title{Function for segmentation}
\usage{
genosegF2(geno, genocols, chromosomes, chrname = "chr", posname = "pos", 
    output.common = "seg", data.type = "logratio", allele1.name = "A", 
    allele2.name = "B", hetero.name = "H", missing.name = c("0"), allele1.code = -1, 
    allele2.code = 1, hetero.code = 0, min.seg.size = 1e+05, cna.alpha = 0.01, 
    cna.nperm = 10000, cna.p.method = "perm", cna.eta = 0.01, cna.min.width = 5, 
    seg.mean.cutoffs = c(-0.8, -0.2, 0.2, 0.8))
}
\arguments{
\item{geno}{genotyping data, containing at least chr; pos; genotyping scores of lines}

\item{genocols}{column numbers of individuals}

\item{chromosomes}{chromosomes or contigs of interest for segmentation}

\item{chrname}{column name/num of chromosomes or contigs}

\item{posname}{column name/num of marker positions}

\item{output.common}{the prefix name of output file}

\item{data.type}{binary or logratio, default logratio}

\item{missing.name}{code for missing data, could be a vector}

\item{min.seg.size}{minimal size of segments, segments with smaller length will be ignored}

\item{seg.mean.cutoffs}{a numeric vector with two values in between 0 and 1, c(smaller, hetero.smaller, hetero.larger, larger),
segment mean values <= smaller: allele 1
segment mean values >= larger: allele 2
segment mean values in between (hetero.smaller, hetero.larger): hetero}

\item{alleleX.name}{allele labels in the output}

\item{alleleX.code}{allele codes as values for segmentation}

\item{cna.xxx}{are DNAcopy parameters, please refer DNAcopy mannual}
}
\value{
output file and variable of the genotyping segmentation result
}
\description{
F2 segmentation using genotyping data
Input data requirment: genotyping data with columns of chr; pos; >=1 lines
The package DNAcopy is required.
}
\examples{
genosegF2(geno = "")

}
\seealso{
\url{https://bioconductor.org/packages/release/bioc/html/DNAcopy.html}
}
\author{
Sanzhen Liu
}
