% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coords.R
\name{get_lg_lengths}
\alias{get_lg_lengths}
\title{Get linkage group extents}
\usage{
get_lg_lengths(obj)
}
\arguments{
\item{obj}{An object of class cross, map or tidy_gen_df}
}
\value{
map coordinates as a vector
}
\description{
Get linkage group extents
}

