% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coords.R
\name{get_lg_names}
\alias{get_lg_names}
\title{Title}
\usage{
get_lg_names(obj)
}
\arguments{
\item{obj}{An object of class cross, map or tidy_gen_df}
}
\value{
A vector of linkage group names
}
\description{
Title
}
