% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_constructors.R
\name{get_lg_stats}
\alias{get_lg_stats}
\title{Returns the maximum distance between markers for a
linkage group.}
\usage{
get_lg_stats(obj)
}
\arguments{
\item{obj}{A cross object suitable for mstmap().}
}
\description{
Returns the maximum distance between markers for a
linkage group.
}

