% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_coords.R
\name{get_map_coords}
\alias{get_map_coords}
\title{Get offsets for linkage groups}
\usage{
get_map_coords(obj, order = NULL)
}
\arguments{
\item{obj}{An object of class cross, map or tidy_gen_df}

\item{order}{A vector of linkage group names to sort map by}
}
\value{
offsets as a vector
}
\description{
Get offsets for linkage groups
}
