% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotters.R
\docType{data}
\name{ggholmans.theme}
\alias{ggholmans.theme}
\title{A theme for trinomial plots}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 4.}
\usage{
ggholmans.theme
}
\description{
A theme for trinomial plots
}
\keyword{datasets}

