% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{isHomo}
\alias{isHomo}
\title{Checks a genotype score for homozygosity}
\usage{
isHomo(score, characterClass = "[[:alnum:]]")
}
\arguments{
\item{score}{A vector of genotype scores, each of which needs to be a two
character string. Numeric values are converted to strings}

\item{characterClass}{A regex character class to match against. If characters
in \code{score} do not exist in the character class, an error will be
returned}
}
\description{
Takes a vector of genotype scores (each made up of a two character string)
and a character class to match against, and returns a logical value depending
on whether the two characters match or not.
}
\examples{
isHomo("AA")
isHomo(c("AA", "AB", "BB"))
isHomo("GT", "[AB]")
}

