% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{join2maps}
\alias{join2maps}
\title{Join two map dfs}
\usage{
join2maps(map1, map2, map1markerName, map2markerName, map1distName,
  map2distName, map1lgName, map2lgName)
}
\arguments{
\item{map1}{The data.frame for the first map.}

\item{map2}{The data.frame for the second map.}

\item{map1markerName}{An unquoted string defining the name for the marker
column in map1.}

\item{map2markerName}{An unquoted string defining the name for the marker
column in map2.}

\item{map1distName}{An unquoted string defining the name for the distance
column in map1.}

\item{map2distName}{An unquoted string defining the name for the distance
column in map2.}

\item{map1lgName}{An unquoted string defining the name for the linkage group
column in map1.}

\item{map2lgName}{An unquoted string defining the name for the linkage group
column in map2.}
}
\description{
Take two maps that have been converted to df (possibly through
/code{map2df()}) and join them.
}
