% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{longmaps}
\alias{longmaps}
\title{Take output from /code{join2maps()} and put into long form for plotting}
\usage{
longmaps(df, reflg_facet, reflg_join, markerName = markerName)
}
\arguments{
\item{df}{A data.frame (output from /code{join2maps()}) of joined map data.}

\item{reflg_facet}{An unquoted string. The name for the reference linkage group identifier for
faceting.}

\item{reflg_join}{An unquoted string. The name for the reference linkage group identifier for
joining.}

\item{markerName}{An unquoted string. The name of the variable giving marker names.}
}
\description{
Refers to a reference linkage group for facetting, and one for joining,
this avoids errors when a marker exists twice in one of the maps.
}

