% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_cross}
\alias{m4_cross}
\title{A multiparental cross object}
\format{A list of class 'mpcross'
\describe{
\item{founders}{The genotyping of the founders}
\item{map}{A list describing the genetic map}
...
}}
\usage{
m4_cross
}
\description{
A 4-way mpcross object produced by mpMap
}
\keyword{datasets}

