% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_int}
\alias{m4_int}
\title{An mpwgaim mpInterval object}
\format{A list of class 'mpInterval'
\describe{
\item{pheno}{A phenotype data.frame}
\item{geno}{A list of lgs, including genotyping and map data}
...
}}
\usage{
m4_int
}
\description{
A 4-way mpInterval object produced by mpwgaim
}
\keyword{datasets}

