% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{make_nptest_classes}
\alias{make_nptest_classes}
\title{Converts a character vector of genotype scores into two classes - one of which represents one homozygous state, and the other represents the other homozygous state and all heterozygotes}
\usage{
make_nptest_classes(vec, level = 1, na_string = "u")
}
\arguments{
\item{vec}{A character vector of genotype scores, each of which is made of two characters. An na_string may also be present}

\item{level}{Either 1 or 2, indicating which homozygous state is to be compared to the other state/heterozygotes. Default is 1}

\item{na_string}{a string representing NA. Default is "u"}
}
\description{
Converts a character vector of genotype scores into two classes - one of which represents one homozygous state, and the other represents the other homozygous state and all heterozygotes
}
