% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{revmap}
\alias{revmap}
\title{Reverse particular linkage groups in a map from a cross object}
\usage{
revmap(map, revgroups)
}
\arguments{
\item{map}{A cross objects}

\item{revgroups}{The linkage groups to be reversed}
}
\description{
Reverse particular linkage groups in a map from a cross object
}
