% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{revmaplgs}
\alias{revmaplgs}
\title{Take output from /code{join2maps()} and a map object to be reversed, and
returns the map object with appropriate groups reversed.}
\usage{
revmaplgs(maps.comp_df, refmapid, revmapid, revmap)
}
\arguments{
\item{maps.comp_df}{Output from /code{join2maps()}.}

\item{refmapid}{The reference map name to identify columns in maps.comp_df.}

\item{revmapid}{The name of the map to be reversed, used to identify columns
in maps.comp_df.}

\item{revmap}{a cross object which is to have linkage groups reversed.}
}
\description{
Take output from /code{join2maps()} and a map object to be reversed, and
returns the map object with appropriate groups reversed.
}

