% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{revvec}
\alias{revvec}
\title{Provide the 'mirror image' of a vector, especially of positions.}
\usage{
revvec(vec)
}
\arguments{
\item{vec}{A (possibly named) numeric vector.}
}
\description{
Provide the 'mirror image' of a vector, especially of positions.
}
\examples{
distvec <- c(A = 0, B = 3, C = 10)
revvec(distvec)
}

