% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg2score.R
\name{seg2score}
\alias{seg2score}
\title{Function for generating common segment markers across all iindividuals}
\usage{
seg2score(seg.input, segscore.output, missing.data.code = NULL, binsize = 1e+06)
}
\arguments{
\item{seg.input}{input data that are the result from seggenoXX moduless}

\item{segscore.output}{output file for genotyping data}

\item{missing.data.code}{missing data code to replace NA. default=NULL.}

\item{binsize}{bin size for facilitating computation, default = 1000000. NOT recommended to change}
}
\value{
genotyping table for individuals
}
\description{
read genotyping segmentation result and divide each segment
into multiple smaller segments. Basically, breakpoints from
all individuals were used to split original segments.
}
\examples{
seg2score(seg.input = "", segscore.output = "")

}
\seealso{
seggenoF2
}
\author{
Sanzhen Liu
}
