% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoplotters.R
\name{trixy}
\alias{trixy}
\title{Convert trinomial distribution into x y values}
\usage{
trixy(df)
}
\arguments{
\item{df}{A data.frame of three columns. Where a row does not sum to 1, it
will be rescaled.}
}
\description{
Takes a data.frame of three columns where each row represents a trinomial
distribution. Returns xy coordinates for plotting in a triangle.\
}

