% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{convertScore}
\alias{convertScore}
\title{Converts a single marker, typically a row in a data.frame.}
\usage{
convertScore(maternal, paternal, progeny, missingString = "--")
}
\arguments{
\item{maternal, paternal}{A two character string, or missingString value.}

\item{progeny}{A (named) character vector of progeny scores.}

\item{missingString}{A string defining missing scores.}
}
\description{
Converts a single marker, typically a row in a data.frame.
}
