% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{convertScore}
\alias{convertScore}
\title{Convert absolute scores to relative}
\usage{
convertScore(maternal, paternal, progeny, markerName = "unknown",
  missingString = "--")
}
\arguments{
\item{parent1}{}
}
\description{
Takes absolute genotype scores and converts to relative scores (AA/AB/BB to
match maternal/het/paternal).
}

