% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoplotters.R
\name{genoPlot}
\alias{genoPlot}
\title{Plot genotype data for a linkage group}
\usage{
genoPlot(df)
}
\arguments{
\item{df}{Sorted data.frame output from /code{genoComp}}
}
\description{
Plot genotype data for a linkage group
}

