% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{genoSeg}
\alias{genoSeg}
\title{Compares segregation of markers based on agreement of two markers (population parents)}
\usage{
genoSeg(df, genos)
}
\description{
Compares segregation of markers based on agreement of two markers (population parents)
}

