% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoplotters.R
\docType{data}
\name{ggholman.theme}
\alias{ggholman.theme}
\title{A theme for trinomial plots}
\format{\preformatted{List of 4
 $ panel.background:List of 4
  ..$ fill    : chr "white"
  ..$ colour  : NULL
  ..$ size    : NULL
  ..$ linetype: NULL
  ..- attr(*, "class")= chr [1:2] "element_rect" "element"
 $ axis.title      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.text       : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 $ axis.ticks      : list()
  ..- attr(*, "class")= chr [1:2] "element_blank" "element"
 - attr(*, "class")= chr [1:2] "theme" "gg"
 - attr(*, "complete")= logi FALSE
}}
\usage{
ggholman.theme
}
\description{
A theme for trinomial plots
}
\keyword{datasets}

