% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{map2df}
\alias{map2df}
\title{Convert R/qtl map into data.frame}
\usage{
map2df(obj)
}
\arguments{
\item{map}{An R/qtl cross or map object}
}
\description{
Similar to pull.map, but returns a more useful object.
}
