% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{map2df}
\alias{map2df}
\title{Convert R/qtl map into data.frame}
\usage{
map2df(map)
}
\arguments{
\item{map}{An R/qtl cross object}
}
\description{
Similar to pull.map, but returns a more useful object.
}

