% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoconverters.R
\name{map2df}
\alias{map2df}
\title{Convert R/qtl map into a tidy data.frame}
\usage{
map2df(obj)
}
\arguments{
\item{obj}{An object of either class cross, map or mpcross}
}
\description{
Similar to pull.map, but returns a tidy data.frame, tibble, tidy_gen_map.
}

